+++ book/tests/prql/queries/dialect_and_version-0.prql
@@ -1,4 +1,4 @@
-prql dialect:postgres
+prql sql_dialect:postgres
 
 from employees
 sort age
