+++ book/tests/prql/queries/dialect_and_version-1.prql
@@ -1,4 +1,4 @@
-prql dialect:mssql
+prql sql_dialect:mssql
 
 from employees
 sort age
