+++ book/tests/prql/queries/syntax-7.prql
@@ -1,3 +1,3 @@
-prql dialect:mysql
+prql sql_dialect:mysql
 from employees
 select `first name`
