+++ book/tests/prql/queries/syntax-8.prql
@@ -1,3 +1,3 @@
-prql dialect:postgres
+prql sql_dialect:postgres
 from employees
 select `first name`
