+++ book/tests/prql/queries/syntax-9.prql
@@ -1,3 +1,3 @@
-prql dialect:bigquery
+prql sql_dialect:bigquery
 from `project-foo.dataset.table`
 join `project-bar.dataset.table` [==col_bax]
