+++ prql-compiler/benches/bench.rs
@@ -17,7 +17,7 @@ cfg_if::cfg_if! {
         use prql_compiler::*;
 
         const CONTENT: &str = include_str!("../../book/tests/prql/examples/variables-0.prql");
-        fn compile_query() -> Result<String> {
+        fn compile_query() -> Result<String, ErrorMessage> {
             compile(CONTENT)
         }
