+++ prql-compiler/src/semantic/mod.rs
@@ -45,7 +45,7 @@ pub fn resolve_only(
 }
 
 pub fn load_std_lib() -> Context {
-    use crate::parse;
+    use crate::parser::parse;
     let std_lib = include_str!("./stdlib.prql");
     let statements = parse(std_lib).unwrap();
 
@@ -72,7 +72,7 @@ mod test {
     use insta::assert_yaml_snapshot;
 
     use super::resolve;
-    use crate::{ast::rq::Query, parse};
+    use crate::{ast::rq::Query, parser::parse};
 
     fn parse_and_resolve(query: &str) -> Result<Query> {
         resolve(parse(query)?)
@@ -81,14 +81,15 @@ mod test {
     #[test]
     fn test_header() {
         assert_yaml_snapshot!(parse_and_resolve(r###"
-        prql dialect:mssql version:"0"
+        prql sql_dialect:mssql version:"0"
 
         from employees
         "###).unwrap(), @r###"
         ---
         def:
           version: ^0
-          dialect: MsSql
+          other:
+            sql_dialect: mssql
         tables:
           - id: 0
             name: employees
@@ -114,14 +115,15 @@ mod test {
         "### );
 
         assert_yaml_snapshot!(parse_and_resolve(r###"
-        prql dialect:bigquery version:"0.3"
+        prql sql_dialect:bigquery version:"0.3"
 
         from employees
         "###).unwrap(), @r###"
         ---
         def:
           version: ^0.3
-          dialect: BigQuery
+          other:
+            sql_dialect: bigquery
         tables:
           - id: 0
             name: employees
@@ -148,7 +150,7 @@ mod test {
 
         assert!(parse_and_resolve(
             r###"
-        prql dialect:bigquery version:foo
+        prql sql_dialect:bigquery version:foo
         from employees
         "###,
         )
@@ -156,7 +158,7 @@ mod test {
 
         assert!(parse_and_resolve(
             r###"
-        prql dialect:bigquery version:"25"
+        prql sql_dialect:bigquery version:"25"
         from employees
         "###,
         )
@@ -164,7 +166,7 @@ mod test {
 
         assert!(parse_and_resolve(
             r###"
-        prql dialect:yah version:foo
+        prql sql_dialect:yah version:foo
         from employees
         "###,
         )
