+++ prql-java/src/lib.rs
@@ -1,7 +1,7 @@
 use jni::objects::{JClass, JString};
 use jni::sys::jstring;
 use jni::JNIEnv;
-use prql_compiler::{parse, pl_to_json};
+use prql_compiler::{json_of_pl, pl_of_prql};
 
 #[no_mangle]
 #[allow(non_snake_case)]
@@ -32,11 +32,8 @@ pub extern "system" fn Java_org_prql_prql4j_PrqlCompiler_toJson(
         .get_string(query)
         .expect("Couldn't get java string!")
         .into();
-    let rs_json_str: String = {
-        let prql: &str = &prql_query;
-        parse(prql).and_then(pl_to_json)
-    }
-    .expect("Couldn't get prql json of query!");
+    let rs_json_str: String =
+        { pl_of_prql(&prql_query).and_then(json_of_pl) }.expect("Couldn't get prql json of query!");
     env.new_string(rs_json_str)
         .expect("Couldn't create java string!")
         .into_raw()
