+++ prql-js/src/lib.rs
@@ -2,128 +2,124 @@
 #![allow(clippy::drop_non_drop)]
 mod utils;
 
-use prql_compiler::{pl_to_json, pl_to_prql, translate, ErrorMessage, IntoErrorMessage};
 use wasm_bindgen::prelude::*;
 
 #[wasm_bindgen]
-pub fn compile(s: &str) -> CompileResult {
-    let result = prql_compiler::compile(s);
-
-    // I had to make new CompileResult struct, because I couldn't make wasm_bindgen
-    // accept it as a function return value. I also had to implement a few getters. Yuck.
-    let mut r = CompileResult::default();
-    match result {
-        Ok(sql) => r.sql = Some(sql),
-        Err(e) => {
-            let error = e.into_error_message("", s, false);
-
-            r.error = Some(CompileError {
-                line: error.line,
-                message: error.message,
-                location: error.location.map(|l| SourceLocation {
-                    start_line: l.start.0,
-                    start_column: l.start.1,
-                    end_line: l.end.0,
-                    end_column: l.end.1,
-                }),
+pub fn compile(prql_query: &str, options: Option<CompileOptions>) -> Option<String> {
+    return_or_throw(
+        Ok(prql_query)
+            .and_then(prql_compiler::pl_of_prql)
+            .and_then(prql_compiler::rq_of_pl)
+            .and_then(|rq| {
+                prql_compiler::sql_of_rq(rq, options.map(prql_compiler::sql::Options::from))
             })
+            .map_err(|e| e.composed("", prql_query, false)),
+    )
+}
+
+#[wasm_bindgen]
+pub fn pl_of_prql(prql_query: &str) -> Option<String> {
+    return_or_throw(
+        Ok(prql_query)
+            .and_then(prql_compiler::pl_of_prql)
+            .and_then(prql_compiler::json_of_pl),
+    )
+}
+
+#[wasm_bindgen]
+pub fn rq_of_pl(pl_json: &str) -> Option<String> {
+    return_or_throw(
+        Ok(pl_json)
+            .and_then(prql_compiler::pl_of_json)
+            .and_then(prql_compiler::rq_of_pl)
+            .and_then(prql_compiler::json_of_rq),
+    )
+}
+
+#[wasm_bindgen]
+pub fn sql_of_rq(rq_json: &str) -> Option<String> {
+    return_or_throw(
+        Ok(rq_json)
+            .and_then(prql_compiler::rq_of_json)
+            .and_then(|x| prql_compiler::sql_of_rq(x, None)),
+    )
+}
+
+/// Compilation options for SQL backend of the compiler.
+#[wasm_bindgen]
+#[derive(Clone)]
+pub struct CompileOptions {
+    /// True for passing generated SQL string trough a formatter that splits
+    /// into multiple lines and prettifies indentation and spacing.
+    pub format: bool,
+
+    /// Target dialect you want to compile for.
+    ///
+    /// Because PRQL compiles to a subset of SQL, not all SQL features are
+    /// required for PRQL. This means that generic dialect may work with most
+    /// databases.
+    ///
+    /// If something does not work in dialect you need, please report it at
+    /// GitHub issues.
+    ///
+    /// If None is used, `sql_dialect` flag from query definition is used.
+    /// If it does not exist, [Dialect::Generic] is used.
+    pub dialect: Option<Dialect>,
+}
+
+#[wasm_bindgen]
+pub fn default_compile_options() -> CompileOptions {
+    CompileOptions {
+        format: true,
+        dialect: None,
+    }
+}
+
+impl From<CompileOptions> for prql_compiler::sql::Options {
+    fn from(o: CompileOptions) -> Self {
+        prql_compiler::sql::Options {
+            format: o.format,
+            dialect: o.dialect.map(From::from),
         }
     }
-    r
 }
 
 #[wasm_bindgen]
-#[derive(Default)]
-pub struct CompileResult {
-    sql: Option<String>,
-    error: Option<CompileError>,
+#[derive(Clone, Copy)]
+pub enum Dialect {
+    Ansi,
+    BigQuery,
+    ClickHouse,
+    Generic,
+    Hive,
+    MsSql,
+    MySql,
+    PostgreSql,
+    SQLite,
+    Snowflake,
 }
 
-#[wasm_bindgen]
-impl CompileResult {
-    #[wasm_bindgen(getter)]
-    pub fn sql(&self) -> Option<String> {
-        self.sql.clone()
-    }
-
-    #[wasm_bindgen(getter)]
-    pub fn error(&self) -> Option<CompileError> {
-        self.error.clone()
-    }
-}
-
-#[wasm_bindgen]
-#[derive(Clone)]
-pub struct CompileError {
-    line: String,
-    message: String,
-    location: Option<SourceLocation>,
-}
-
-#[wasm_bindgen]
-impl CompileError {
-    #[wasm_bindgen(getter)]
-    pub fn line(&self) -> String {
-        self.line.clone()
-    }
-
-    #[wasm_bindgen(getter)]
-    pub fn message(&self) -> String {
-        self.message.clone()
-    }
-
-    #[wasm_bindgen(getter)]
-    pub fn location(&self) -> Option<SourceLocation> {
-        self.location.clone()
+impl From<Dialect> for prql_compiler::sql::Dialect {
+    fn from(d: Dialect) -> Self {
+        use prql_compiler::sql::Dialect as D;
+        match d {
+            Dialect::Ansi => D::Ansi,
+            Dialect::BigQuery => D::BigQuery,
+            Dialect::ClickHouse => D::ClickHouse,
+            Dialect::Generic => D::Generic,
+            Dialect::Hive => D::Hive,
+            Dialect::MsSql => D::MsSql,
+            Dialect::MySql => D::MySql,
+            Dialect::PostgreSql => D::PostgreSql,
+            Dialect::SQLite => D::SQLite,
+            Dialect::Snowflake => D::Snowflake,
+        }
     }
 }
 
-// TODO: to what extent should this be unified with the SourecLocation in error.rs?
-#[wasm_bindgen]
-#[derive(Clone)]
-pub struct SourceLocation {
-    pub start_line: usize,
-    pub start_column: usize,
-
-    pub end_line: usize,
-    pub end_column: usize,
-}
-
-#[wasm_bindgen]
-pub fn to_sql(s: &str) -> Option<String> {
-    let result = prql_compiler::compile(s).map_err(|e| e.into_error_message("", s, false));
-    return_or_throw_error(result)
-}
-
-#[wasm_bindgen]
-pub fn to_json(s: &str) -> Option<String> {
-    let result = prql_compiler::parse(s)
-        .and_then(pl_to_json)
-        .map_err(|e| e.into_error_message("", s, false));
-    return_or_throw_error(result)
-}
-
-#[wasm_bindgen]
-pub fn from_json(s: &str) -> Option<String> {
-    let result = prql_compiler::json_to_pl(s)
-        .and_then(pl_to_prql)
-        .map_err(|e| e.into_error_message("", s, false));
-    return_or_throw_error(result)
-}
-
-#[wasm_bindgen]
-pub fn json_to_rq_to_sql(s: &str) -> Option<String> {
-    let result = prql_compiler::json_to_rq(s)
-        .and_then(translate)
-        .map_err(|e| e.into_error_message("", s, false));
-    return_or_throw_error(result)
-}
-
-fn return_or_throw_error(result: Result<String, ErrorMessage>) -> Option<String> {
+fn return_or_throw(result: Result<String, prql_compiler::ErrorMessages>) -> Option<String> {
     match result {
         Ok(sql) => Some(sql),
-        Err(e) => {
-            wasm_bindgen::throw_str(&e.message);
-        }
+        Err(e) => wasm_bindgen::throw_str(&e.to_json()),
     }
 }
