+++ prql-macros/src/lib.rs
@@ -1,5 +1,4 @@
 use proc_macro::{Literal, TokenStream, TokenTree};
-use prql_compiler::{compile, IntoErrorMessage};
 use syn::{Expr, ExprLit, Lit};
 
 #[proc_macro]
@@ -14,11 +13,10 @@ pub fn prql(input: TokenStream) -> TokenStream {
         _ => panic!("prql! proc macro expected a string"),
     };
 
-    let sql_string = match compile(&prql_string) {
+    let sql_string = match prql_compiler::compile(&prql_string) {
         Ok(r) => r,
         Err(err) => {
-            let err = err.into_error_message("<prql_macro>", &prql_string, true);
-            panic!("{}", err.message);
+            panic!("{}", err);
         }
     };
