+++ prql-python/README.md
@@ -1,7 +1,7 @@
 # prql-python
 
 `prql-python` offers rust bindings to the `prql-compiler` rust library. It
-exposes a python method `to_sql(query: str) -> str`.
+exposes a python method `compile(query: str) -> str`.
 
 This is consumed by [PyPrql](https://github.com/prql/PyPrql) &
 [dbt-prql](https://github.com/prql/dbt-prql).
@@ -27,13 +27,7 @@ prql_query = """
     )
 """
 
-sql = prql.to_sql(prql_query)
+sql = prql.compile(prql_query)
 ```
 
 Relies on [pyo3](https://github.com/PyO3/pyo3) for all the magic.
-
-```rust
-#[pyfunction]
-pub fn to_sql(query: &str) -> PyResult<String> {}
-fn prql_python(_py: Python, m: &PyModule) -> PyResult<()> {}
-```
