+++ prql-python/python/tests/test_all.py
@@ -11,10 +11,13 @@ def test_all():
     prql_query = "from employee"
 
     # Since the AST is so in flux, let's just take these dont throw exceptions
-    res = json.loads(prql.to_json(prql_query))
+    res = prql.pl_of_prql(prql_query)
     assert res is not None
 
-    res = prql.to_sql(prql_query)
+    res = prql.rq_of_pl(res)
+    assert res is not None
+
+    res = prql.sql_of_rq(res)
     assert res is not None
 
     assert prql.__version__ is not None
